requirements.txt
python-telegram-bot==5.1.1
beautifulsoup4==4.5.1
urllib3==1.18
